Fixed Forms
Fixed stored procedures on most all forms except 1 on the login form
where it checks if the user is an admin. It had the user always being an
admin(I think the stored procedure code needs to be tweaked?)

Also need to figure out how to populate the text boxes on the edituser
form.